+++ CHANGELOG.md
@@ -37,10 +37,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## Unreleased
 
+### State Machine Breaking
+
+- [#1826](https://github.com/babylonlabs-io/babylon/pull/1826) Support M-of-N multisig btc staker
+
 ### Improvements
 
 - [#1878](https://github.com/babylonlabs-io/babylon/pull/1878) Remove unnecessary []byte allocations in IBC size checks
-- [#1826](https://github.com/babylonlabs-io/babylon/pull/1826) Support M-of-N multisig btc staker
 
 ### Bug Fixes